Split user list views into multiple files
